[-]修改了已经弃用的kotlinx，改用viewBinding [+]修改了java代码，优化了部分逻辑的实现

+++ app/src/main/java/com/example/petscoffee/repository/local/CoffeeShopDao.kt
@@ -1,9 +1,8 @@
 package com.example.petscoffee.repository.local
 
-import com.example.petscoffee.model.CoffeeShop
 import androidx.lifecycle.LiveData
 import androidx.room.*
-import com.example.petscoffee.model.pets.Pets
+import com.example.petscoffee.model.CoffeeShop
 
 @Dao
 interface CoffeeShopDao {
@@ -11,13 +10,13 @@ interface CoffeeShopDao {
     fun insertCoffee(coffee: CoffeeShop)
 
     @Query("SELECT * FROM coffeeShop WHERE id =:id")
-    fun queryCoffee(id: Int): CoffeeShop
+    fun queryCoffee(id: Long): CoffeeShop
 
     @Query("SELECT * FROM coffeeShop WHERE account=:account")
     fun queryCoffee(account: String): CoffeeShop
 
     @Query("SELECT * FROM coffeeShop WHERE id=:id")
-    fun queryCoffeeLiveData(id: Int): LiveData<CoffeeShop>
+    fun queryCoffeeLiveData(id: Long): LiveData<CoffeeShop>
 
     @Delete
     fun deleteCoffee(coffee: CoffeeShop)
